+++ tests/testthat/test-exec.R
@@ -135,3 +135,8 @@ testthat::test_that("`strelka()` works as expected", {
   testthat::skip_if_not(nzchar(Sys.which("strelka")))
   strelka() |> cmd_help()
 })
+
+testthat::test_that("`minimap2()` works as expected", {
+  testthat::skip_if_not(nzchar(Sys.which("strelka")))
+  strelka() |> cmd_help()
+})